Grain dropdown fetched from the DB. Fixed pH name bug.

+++ src/main/java/hr/brewer/models/DecreasePhAcid.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 public @Data class DecreasePhAcid {
@@ -12,7 +13,7 @@ public @Data class DecreasePhAcid {
     @Id
     @GeneratedValue
     private Long waterId;
-    private Integer lacticAcid;
-    private Integer acidulatedMalt;
+    private BigDecimal lacticAcid;
+    private BigDecimal acidulatedMalt;
 
 }
