+++ src/main/java/hr/brewer/models/DecreasePhSalts.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 public @Data class DecreasePhSalts {
@@ -12,9 +13,9 @@ public @Data class DecreasePhSalts {
     @Id
     @GeneratedValue
     private Long waterId;
-    private Integer epsomSalt;
-    private Integer calciumChloride;
-    private Integer gypsum;
+    private BigDecimal epsomSalt;
+    private BigDecimal calciumChloride;
+    private BigDecimal gypsum;
     private boolean showGypsum;
     private boolean showCalciumChloride;
     private boolean showEpsomSalt;
