+++ src/main/java/hr/brewer/models/GrainBill.java
@@ -3,6 +3,7 @@ package hr.brewer.models;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -13,7 +14,7 @@ public @Data class GrainBill {
     private Long waterId;
     @OneToMany(targetEntity=Grain.class, cascade=CascadeType.ALL)
     private List<Grain> grains;
-    private Integer mashThickness;
-    private Integer totalGrainWeight;
+    private BigDecimal mashThickness;
+    private BigDecimal totalGrainWeight;
 
 }
